requirements.txt
<<<<<<< HEAD
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
colorama==0.4.6
distro==1.9.0
fastavro==1.11.1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
openai==1.82.0
packaging==25.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250515
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
=======
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
colorama==0.4.6
distro==1.9.0
fastavro==1.11.1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
openai==1.82.0
packaging==25.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250515
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
>>>>>>> 92da4fbc469d49fa37f98288f1e2b73024a7521b
